# Validation optimizations: Tasks, List/iterators, Action/Func closures, etc