``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.16299.726 (1709/FallCreatorsUpdate/Redstone3)
Intel Core i7 CPU 920 2.67GHz (Nehalem), 1 CPU, 8 logical and 4 physical cores
Frequency=2607421 Hz, Resolution=383.5207 ns, Timer=TSC
.NET Core SDK=5.0.101
  [Host]     : .NET Core 3.1.10 (CoreCLR 4.700.20.51601, CoreFX 4.700.20.51901), X64 RyuJIT
  DefaultJob : .NET Core 3.1.10 (CoreCLR 4.700.20.51601, CoreFX 4.700.20.51901), X64 RyuJIT

```
before/after (before from #2156)

|        Method |      Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------- |----------:|---------:|---------:|-------:|------:|------:|----------:|
| Introspection | 426.06 μs | 0.588 μs | 0.491 μs | 3.4180 |     - |     - |  14.16 KB |
|     Fragments | 193.79 μs | 0.123 μs | 0.115 μs | 2.4414 |     - |     - |  10.07 KB |
|          Hero |  29.66 μs | 0.055 μs | 0.051 μs | 1.0681 |     - |     - |   4.48 KB |

|        Method |      Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------- |----------:|---------:|---------:|-------:|------:|------:|----------:|
| Introspection | 420.03 μs | 0.195 μs | 0.173 μs | 2.4414 |     - |     - |  11.53 KB |
|     Fragments | 189.38 μs | 0.119 μs | 0.111 μs | 1.7090 |     - |     - |   7.95 KB |
|          Hero |  25.92 μs | 0.028 μs | 0.024 μs | 0.7019 |     - |     - |   2.91 KB |

My main goal (again) was to reduce the number of heap allocations. It's nice that the speed has also increased.

[UPDATE] Looking at our old PRs, I saw that the memory consumption for `Hero` case decreased by more than 5 times.